+++ src/main/data/duke.txt
@@ -1,0 +1,2 @@
+1. [T][ ] say hi
+2. [D][ ] lzdkf (by:Dec 24 2022)
