+++ src/main/java/duke/Event.java
@@ -1,6 +1,6 @@
+package duke;
+
 import exceptions.DukeEventException;
-import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
 
 public class Event extends Task {
@@ -22,8 +22,6 @@ public class Event extends Task {
             eventname = split[0];
             eventat = split[1];
             Event e = new Event(eventname, eventat);
-//            System.out.println("Got it. I've added this task: ");
-//            System.out.println(e.toString());
             return e;
         } catch (Exception e) {
             DukeEventException error = new DukeEventException(
