+++ src/main/java/duke/ListCommand.java
@@ -1,6 +1,8 @@
+package duke;
 public class ListCommand extends Command{
 
     public String execute(TaskList tasklist, Ui ui, Storage storage) {
+
         return ui.printAllTasks(tasklist);
     }
 }
