+++ src/main/java/duke/MarkCommand.java
@@ -1,3 +1,4 @@
+package duke;
 import java.io.IOException;
 
 /**
