+++ src/main/java/duke/Ui.java
@@ -1,4 +1,4 @@
-import exceptions.DukeException;
+package duke;
 
 import java.util.ArrayList;
 import java.util.Scanner;
